fix(lint): import order, unused import, react-refresh warning (CI)
[Agent: Main Assistant]

Fixes the two failing PR #302 checks:
- Python Lint: worker.py import block re-sorted (I001); unused
  pathlib.Path dropped from tests/test_transcription_review.py (F401)
- Frontend Lint: formatClock moved from SegmentRow.tsx to
  utils/formatTime.ts — component files must only export components
  (react-refresh/only-export-components, --max-warnings 0)

ruff check ., eslint --max-warnings 0, tsc, and the affected test files
all pass locally.
