+++ web/src/components/review/SegmentRow.tsx
@@ -1,4 +1,5 @@
 import { memo, useLayoutEffect, useRef } from 'react'
+import { formatClock } from '../../utils/formatTime'
 
 export interface EditableSegment {
   id: number
@@ -6,16 +7,6 @@ export interface EditableSegment {
   end: number
   speaker: string | null
   text: string
-}
-
-export function formatClock(seconds: number): string {
-  const total = Math.max(0, Math.floor(seconds))
-  const h = Math.floor(total / 3600)
-  const m = Math.floor((total % 3600) / 60)
-  const s = total % 60
-  return h > 0
-    ? `${h}:${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`
-    : `${m}:${String(s).padStart(2, '0')}`
 }
 
 interface SegmentRowProps {
